requirements.txt
asgiref==3.9.1
Brotli==1.2.0
cffi==1.17.1
cssselect2==0.8.0
fonttools==4.60.2
Pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
Unidecode==1.3.8
webencodings==0.5.1
zopfli==0.2.3.post1

daphne>=4.0

# Django and related
Django>=4.2,<5.0
channels==4.1.0
channels-redis==4.2.0
uvicorn==0.30.6
gunicorn
whitenoise
dj-database-url==3.0.1

# Rate limiting
django-ratelimit==4.1.0

# REST API
djangorestframework>=3.16,<4.0

# Cloudflare R2 (S3-compatible)
django-storages
boto3

# Redis & Caching
django-redis==5.4.0
redis==5.0.1

# Excel handling
openpyxl==3.1.5

# PDF
weasyprint

# MIME detection (optional; validators fall back if unavailable)
python-magic==0.4.27

# Celery
celery==5.4.0
django-celery-results==2.5.1

# Markdown rendering (User guide page)
Markdown==3.7
